requirements.txt
Django>=2.2.0, <2.3.0
djangorestframework>=3.12.2, <3.13.0
dj-rest-auth>=2.1.3, <2.2.0
allauth-django>=0.42.6, <0.43.0
django-cors-headers==3.7.0
pytest-django>=4.1.0, <4.2.0
flake8>=3.8.4, <3.9.0
psycopg2-binary>=2.8.6, <2.9.0
Pillow>=5.3.0, <5.4.0
